+++ .storybook/preview.tsx
@@ -14,7 +14,7 @@ import {
   Link,
   LinkProps,
   List,
-  ListItem,
+  Table,
 } from '@digdir/design-system-react';
 import customTheme from './customTheme';
 import metadata from '../design-tokens/$metadata.json';
@@ -59,7 +59,13 @@ const getPath = (href: string | undefined): string => {
 };
 
 const components = {
-  p: Paragraph,
+  p: (props: Props) => (
+    <Paragraph
+      {...props}
+      className='sb-unstyled'
+      spacing
+    />
+  ),
   ol: (props: Props) => (
     <List.Root>
       <List.Ordered
@@ -93,9 +99,48 @@ const components = {
       <Link
         {...props}
         href={href}
+        className='sb-unstyled'
       ></Link>
     );
   },
+  table: (props: Props) => (
+    <Table
+      {...props}
+      zebra
+      className='sb-unstyled'
+      style={{ width: '100%' }}
+    />
+  ),
+  thead: (props: Props) => (
+    <Table.Head
+      {...props}
+      className='sb-unstyled'
+    />
+  ),
+  tbody: (props: Props) => (
+    <Table.Body
+      {...props}
+      className='sb-unstyled'
+    />
+  ),
+  tr: (props: Props) => (
+    <Table.Row
+      {...props}
+      className='sb-unstyled'
+    />
+  ),
+  th: (props: Props) => (
+    <Table.HeaderCell
+      {...props}
+      className='sb-unstyled'
+    />
+  ),
+  td: (props: Props) => (
+    <Table.Cell
+      {...props}
+      className='sb-unstyled'
+    />
+  ),
 };
 
 const preview: Preview = {
@@ -132,7 +177,6 @@ const preview: Preview = {
           'Felles',
           'Experimental',
           'Primitives',
-          'Altinn',
           'Avviklet',
         ],
       },