+++ apps/storefront/components/Blog/index.ts
@@ -1,0 +1,2 @@
+export { BlogCard } from './Card/BlogCard';
+export { Contributors } from './Contributors/Contributors';