+++ packages/react/src/components/DropdownMenu/DropdownMenuTrigger.tsx
@@ -1,4 +1,3 @@
-import { Slot } from '@radix-ui/react-slot';
 import { forwardRef, useContext } from 'react';
 import type * as React from 'react';
 import { useMergeRefs } from '@floating-ui/react';
@@ -7,22 +6,20 @@ import { Button } from '../Button';
 
 import { DropdownMenuContext } from './DropdownMenu';
 
-export type DropdownMenuTriggerProps = {
-  asChild?: boolean;
-} & React.ComponentPropsWithRef<typeof Button>;
+export type DropdownMenuTriggerProps = React.ComponentPropsWithRef<
+  typeof Button
+>;
 
 export const DropdownMenuTrigger = forwardRef<
   HTMLButtonElement,
   DropdownMenuTriggerProps
->(({ asChild, children, ...rest }, ref) => {
-  const Component = asChild ? Slot : Button;
-
+>(({ ...rest }, ref) => {
   const { triggerRef, internalOpen, setInternalOpen, isControlled } =
     useContext(DropdownMenuContext);
   const mergedRefs = useMergeRefs([ref, triggerRef]);
 
   return (
-    <Component
+    <Button
       ref={mergedRefs}
       onClick={() => {
         if (!isControlled) setInternalOpen(!internalOpen);
@@ -30,8 +27,8 @@ export const DropdownMenuTrigger = forwardRef<
       aria-haspopup='menu'
       aria-expanded={internalOpen}
       {...rest}
-    >
-      {children}
-    </Component>
+    />
   );
 });
+
+DropdownMenuTrigger.displayName = 'DropdownMenuTrigger';