+++ packages/react/src/components/Modal/ModalFooter/ModalFooter.tsx
@@ -1,19 +1,30 @@
 import type { HTMLAttributes } from 'react';
 import { forwardRef } from 'react';
 import cl from 'clsx';
+import { Slot } from '@radix-ui/react-slot';
 
 import classes from './ModalFooter.module.css';
 
-export const ModalFooter = forwardRef<HTMLElement, HTMLAttributes<HTMLElement>>(
-  ({ children, className, ...rest }, ref) => {
+export type ModalFooterProps = {
+  /**
+   * Change the default rendered element for the one passed as a child, merging their props and behavior.
+   * @default false
+   */
+  asChild?: boolean;
+} & HTMLAttributes<HTMLElement>;
+
+export const ModalFooter = forwardRef<HTMLElement, ModalFooterProps>(
+  ({ asChild, className, ...rest }, ref) => {
+    const Component = asChild ? Slot : 'footer';
+
     return (
-      <footer
+      <Component
         ref={ref}
         className={cl(classes.modalFooter, className)}
         {...rest}
-      >
-        {children}
-      </footer>
+      />
     );
   },
 );
+
+ModalFooter.displayName = 'ModalFooter';